refactor(itinerary): abstract itinerary component
