+++ ui/src/routes/Itinerary.jsx
@@ -10,47 +10,10 @@ import {
 import {
   arrayMove,
   SortableContext,
-  useSortable,
   verticalListSortingStrategy,
 } from "@dnd-kit/sortable";
-import { CSS } from "@dnd-kit/utilities";
 import PageTemplate from "../components/PageTemplate";
-
-const SortableItem = ({ id, place, time, transport, image }) => {
-  const { attributes, listeners, setNodeRef, transform } = useSortable({ id });
-
-  const style = {
-    transform: transform ? CSS.Transform.toString(transform) : undefined,
-    transition: "none",
-  };
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      {...attributes}
-      {...listeners}
-      className="shadow-primary/20 rounded-lg p-3 mb-4 cursor-grab bg-white shadow-md"
-    >
-      <div className="flex items-start">
-        <img
-          src={image}
-          alt={place}
-          className="w-24 h-24 object-cover rounded mr-4"
-        />
-        <div className="flex-1">
-          <h3 className="font-semibold">{place}</h3>
-          <p className="text-sm text-gray-500">{time}</p>
-          {transport && (
-            <p className="text-xs text-gray-400">
-              {transport.type} - {transport.duration}
-            </p>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
+import SortableItem from "../components/SortableItem"; 
 
 function Itinerary() {
   const [itinerary, setItinerary] = useState({});
